+++ game/pieces/__init__.py
@@ -1,0 +1,6 @@
+from .rook import Rook
+from .pawn import Pawn
+from .bishop import Bishop
+from .king import King
+from .queen import Queen
+from .horse import Horse