+++ tests/test_board.py
@@ -1,8 +1,9 @@
 import unittest
-from game.rook import Rook , Pawn
+from game.pieces import Rook , Pawn , Bishop , King , Queen , Horse
 from game.board import Board
 
 class TestBoard(unittest.TestCase):
+    
     def test_init(self):
         board = Board()
         self.assertIsNone(board.get_piece(3, 1))
@@ -32,13 +33,7 @@ class TestBoard(unittest.TestCase):
         piece = board.get_piece(7, 7)
         self.assertEqual(piece.__color__, 'WHITE')
 
-    def test_get_piece_name(self):
-        board = Board()
-        piece = board.get_piece(0, 0)
-        self.assertEqual(piece.__name__, 'ROOK')
 
-        piece = board.get_piece(1, 0)
-        self.assertEqual(piece.__name__, 'PAWN')
 
 
 if __name__ == '__main__':